+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - set cauldron gift to a fixed amount
 - show game rules before joining
 - track player's last activity and don't send messages to the player after 30 days of inactivity
+- add attack and defense ranges
 
 ## v0.1.0